feat: add pit feature, configure option
