+++ .gitignore
@@ -15,6 +15,10 @@ demo_debug
 
 # ignore basic example binary
 basic_example
+# ignore pit_example binary
+pit_example
+# ignore list_example binary
+list_example
 
 #We also want to ignore the dotfile dump if we ever use anvil with -c flag
 amboso_cfg.dot
